moving flask server code to test
