--- backend/pluto/tests/_internal/server/test_flask_server.py
@@ -1,1 +1,0 @@
-import pytest
